// types.ts
export interface Message {
    id: string;
    sender: string;
    text: string;
    timestamp: string;
    isMe: boolean;
    avatar: string;
}